Fix a selection bug and refactor some codes
