+++ src/hooks/filterHook/useUpdateRoutesWithFilter.ts
@@ -1,14 +1,20 @@
-import { useEffect } from 'react';
+import mapboxgl from 'mapbox-gl';
+import React, { useEffect } from 'react';
 
 import {
     selectPTRoutesFeatureList,
     updateFilteredRoutes,
     updateSelectedRoute,
     updateSelectedStop,
+    updateVisibleRouteState,
 } from '../../dataStoring/slice';
 import { useAppDispatch, useAppSelector } from '../../store';
+import { getVisibleRoutes } from './useVisibleRoutesUpdate';
 
-export const useUpdateRoutesWithFilter = () => {
+export const useUpdateRoutesWithFilter = (
+    map: mapboxgl.Map | null,
+    setMap: React.Dispatch<React.SetStateAction<mapboxgl.Map | null>>,
+) => {
     const dispatch = useAppDispatch();
     const filters = useAppSelector((state) => state.slice.filters);
     const routes = useAppSelector(selectPTRoutesFeatureList);
@@ -45,6 +51,25 @@ export const useUpdateRoutesWithFilter = () => {
 
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [filters, routes]);
+
+    const visibleRoutes = useAppSelector((state) => state.slice.visibleRoutes);
+
+    // Update the visible routes when map is moved.
+    useEffect(() => {
+        if (!map || !visibleRoutes.isOn) return;
+        const listener = () => {
+            setMap(newMap);
+            const updatedVisibleFiltering = getVisibleRoutes(map, visibleRoutes);
+            dispatch(updateVisibleRouteState(updatedVisibleFiltering));
+        };
+
+        const newMap = map.on('moveend', listener);
+
+        return () => {
+            map.off('moveend', listener);
+        };
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [map, dispatch, visibleRoutes]);
 };
 
 /**
